Add helper functions `trackUploadProgress` and `removeUploadProgress` for `WoltlabCoreFileElement`.
